fix(tail): Restore cursor after empty 'tail' session
When running `jhack tail` and keyboard-interrupting the
session without any events observed, `live.stop()` will
never be called, resulting in the terminal cursor state
to get corrupted (cursor is hidden in shell).

Fix this by moving the call to `live.stop()` to the
`finally` block, so that it will always be called.