chore(configure): Remove duplicate call
`configure()` is called in `__init__.py` before `main()`, however
`main()` calls `configure()` before doing anything else.

I'm removing the call in `__init__.py`, since the history shows that the
second call in `main.py` was added after that and I believe the intent
was that it is called there (and it was not cleaned up).